hardlink.py: Prevent overwriting already existing backup files (Closes: #514772)
From now on, change the suffix of the created backup files to be .hardlink-PID,
whereas PID is replaced by the ID of the hardlink process. This reduces naming
collisions with already existing files.

Furthermore, if this file already exists, the file linking would be aborted
with an error message (other files will still be linked).

+++ hardlink.py
@@ -99,7 +99,10 @@ class File(object):
 
     def link(self, other):
         '''Link the current file to another one'''
-        backup = other.path + '.bak'
+        backup = other.path + '.hardlink-%s' % os.getpid()
+        if os.path.exists(backup):
+            print 'E: Backup file %s already exists, aborting.' % backup
+            return
         try:
             if not self.opts.dry_run:
                 os.rename(other.path, backup)
